add script para habilitar acesso ao solr somente para ips internos

+++ docker/dspace/scripts/entry-point-dspace.sh
@@ -137,10 +137,17 @@ function habilita_solr_para_acesso_remoto() {
 function habilitar_acesso_ao_solr_somente_ips_internos() {
   arquivo_server_xml="/opt/apache-tomcat-8.5.51/conf/server.xml"
   codigo_xml="<Context path=\"/solr\" reloadable=\"true\">\n        <Valve className=\"org.apache.catalina.valves.RemoteAddrValve\" allow=\"127\.0\.0\.1|172\.16\.16\.d+\"/>\n        <Parameter name=\"LocalHostRestrictionFilter.localhost\" value=\"false\" override=\"false\" />\n</Context>"
+  
+  if [ ! -f "$arquivo_server_xml" ]; then
+    echo "Erro: O arquivo $arquivo_server_xml não existe."
+    return 1
+  fi
+
   if grep -q "$codigo_xml" "$arquivo_server_xml"; then
     echo "O código XML já está presente no arquivo $arquivo_server_xml."
     return 1
   fi
+  
   cp -f "$arquivo_server_xml" "$arquivo_server_xml.bak"  
   echo "$codigo_xml" >> "$arquivo_server_xml"
   echo "XML code added (or overwritten) to $arquivo_server_xml."
@@ -156,7 +163,6 @@ function habilita_debug_remoto() {
 function verfica_e_trata_ambiente_de_desenvolvimento() {
   if [ -n "${ENVIRONMENT}" ]; then
     habilita_solr_para_acesso_remoto
-    habilitar_acesso_ao_solr_somente_ips_internos
     habilita_debug_remoto
   fi
 }
@@ -187,6 +193,7 @@ if [[ ! -f "/opt/docker-build-complete" ]]; then
   remove_arquivos_instalacao
   cria_arquivo_indicador_conclusao_build
   verfica_e_trata_ambiente_de_desenvolvimento
+  habilitar_acesso_ao_solr_somente_ips_internos
 else
   prepara_ambiente_rede
   prepara_tomcat
@@ -196,6 +203,7 @@ else
   instala_dspace_ou_atualiza_dspace
   remove_arquivos_instalacao
   verfica_e_trata_ambiente_de_desenvolvimento
+  habilitar_acesso_ao_solr_somente_ips_internos
 fi
 
 inicia_servicos
